+++ apps/vscode/src/test/utils/event.ts
